+++ tests/fixtures/packages/features/src/lib.rs
@@ -1,0 +1,6 @@
+//! <!-- SYNC_RDME_INTEGRATION_TEST::SPAN_START -->
+#![cfg_attr(feature = "default", doc = "* DEFAULT")]
+#![cfg_attr(feature = "feat-a", doc = "* FEAT_A")]
+#![cfg_attr(feature = "feat-b", doc = "* FEAT_B")]
+#![cfg_attr(feature = "feat-c", doc = "* FEAT_C")]
+//! <!-- SYNC_RDME_INTEGRATION_TEST::SPAN_END -->